refactor: drop int() casts around round()
Changed:
Removed `int()` casts around `round()` calls, which already return an int.

Benefit:
The rounding intent is not hidden behind a redundant cast, enforced by ruff.

+++ src/api/flaskr/api/tts/tencent_provider.py
@@ -278,7 +278,7 @@ def _tencent_pcm_duration_ms(audio_data: bytes, *, sample_rate: int) -> int:
     if not audio_data:
         return 0
     bytes_per_second = max(int(sample_rate or TENCENT_DEFAULT_SAMPLE_RATE), 1) * 2
-    return int(round(len(audio_data) * 1000 / bytes_per_second))
+    return round(len(audio_data) * 1000 / bytes_per_second)
 
 
 def _export_tencent_pcm_to_mp3(audio_data: bytes, *, sample_rate: int) -> bytes:
@@ -689,7 +689,7 @@ def _group_tencent_subtitle_cues_by_source_text(
     def _time_at_weight(target_weight: float) -> int:
         if total_cue_weight <= 0:
             ratio = target_weight / total_sentence_weight
-            return start_ms + int(round(duration_ms * ratio))
+            return start_ms + round(duration_ms * ratio)
 
         bounded_target = max(min(float(target_weight), float(total_cue_weight)), 0.0)
         consumed = 0.0
@@ -705,9 +705,7 @@ def _group_tencent_subtitle_cues_by_source_text(
                 continue
             if bounded_target <= consumed + weight:
                 ratio = (bounded_target - consumed) / weight
-                target_ms = cue_start_ms + int(
-                    round((cue_end_ms - cue_start_ms) * ratio)
-                )
+                target_ms = cue_start_ms + round((cue_end_ms - cue_start_ms) * ratio)
                 if bounded_target >= consumed + weight:
                     for next_cue, next_weight in zip(
                         normalized_cues[index + 1 :],
@@ -743,11 +741,11 @@ def _group_tencent_subtitle_cues_by_source_text(
             cue_start_ms = _time_at_weight(source_start_weight)
             cue_end_ms = _time_at_weight(source_end_weight)
         else:
-            cue_start_ms = start_ms + int(
-                round(duration_ms * sentence_cursor / total_sentence_weight)
+            cue_start_ms = start_ms + round(
+                duration_ms * sentence_cursor / total_sentence_weight
             )
-            cue_end_ms = start_ms + int(
-                round(duration_ms * next_sentence_cursor / total_sentence_weight)
+            cue_end_ms = start_ms + round(
+                duration_ms * next_sentence_cursor / total_sentence_weight
             )
 
         cue_start_ms = max(cue_start_ms, timeline_cursor_ms)
@@ -787,7 +785,7 @@ def _tencent_subtitle_time_ms(
         if value is None or value == "":
             continue
         try:
-            return int(round(float(value)))
+            return round(float(value))
         except (TypeError, ValueError):
             continue
     return int(default_ms or 0)
